+++ tests/testcases/core/domain/entities/routing/specifications/RouteMatchSpecificationDecoratorTest.php
@@ -1,49 +1,51 @@
 <?php
 
-namespace EventEspresso\tests\testcases\core\domain\entities\routing;
+namespace EventEspresso\tests\testcases\core\domain\entities\routing\specifications;
 
-use EventEspresso\core\domain\entities\routing\specifications\DoesNotMatchRouteSpecification;
 use EventEspresso\core\domain\entities\routing\specifications\RouteMatchSpecificationInterface;
+use EventEspresso\tests\mocks\core\domain\entities\routing\specifications\RouteMatchSpecificationDecoratorMock;
+use PHPUnit\Framework\AssertionFailedError;
+use PHPUnit\Framework\Exception;
 
 /**
- * Class DoesNotMatchRouteSpecificationTest
+ * Class RouteMatchSpecificationDecoratorTest
  * Description
  *
  * @package EventEspresso\tests\testcases\core\domain\entities\routing
  * @author  Brent Christensen
  * @since   4.9.71.p
  */
-class DoesNotMatchRouteSpecificationTest extends RouteMatchSpecificationDecoratorTest
+class RouteMatchSpecificationDecoratorTest extends RouteMatchSpecificationTestBase
 {
     /**
-     * @param \EventEspresso\core\domain\entities\routing\specifications\RouteMatchSpecificationInterface|null $route_match_specification
-     * @return DoesNotMatchRouteSpecification
-     *@since 4.9.71.p
+     * @since 4.9.71.p
+     * @param RouteMatchSpecificationInterface|null $route_match_specification
+     * @return RouteMatchSpecificationDecoratorMock
      */
     public function getDecorator(RouteMatchSpecificationInterface $route_match_specification = null)
     {
         $route_match_specification = $route_match_specification instanceof RouteMatchSpecificationInterface
             ? $route_match_specification
             : $this->getRouteMatchSpecification();
-        return new DoesNotMatchRouteSpecification($route_match_specification);
+        return new RouteMatchSpecificationDecoratorMock($route_match_specification);
     }
 
     /**
      * @since 4.9.71.p
-     * @throws \PHPUnit\Framework\Exception
+     * @throws Exception
      */
     public function test__construct()
     {
-        // useless test but need to override parent constructor
         $this->assertInstanceOf(
-            'EventEspresso\core\domain\entities\routing\DoesNotMatchRouteSpecification',
-            $this->getDecorator()
+            'EventEspresso\tests\mocks\core\domain\entities\routing\specifications\RouteMatchSpecificationMock',
+            $this->getDecorator()->getSpecification()
         );
 
     }
+
     /**
      * @since 4.9.71.p
-     * @throws \PHPUnit\Framework\AssertionFailedError
+     * @throws AssertionFailedError
      */
     public function testIsMatchingRoute()
     {
@@ -52,21 +54,18 @@ class DoesNotMatchRouteSpecificationTest extends RouteMatchSpecificationDecorato
         $route_match_specification->setParam('testing');
         $route_match_specification->setValue('one-two-one-two');
         $decorator = $this->getDecorator($route_match_specification);
-        // IT'S OPPOSITE DAY!!!
-        $this->assertTrue($decorator->isMatchingRoute());
+        $this->assertFalse($decorator->isMatchingRoute());
         //  testing 1-2-1-2 request but checking for other route
         $route_match_specification = $this->getRouteMatchSpecification(array('testing' => 'one-two-one-two'));
         $route_match_specification->setParam('testing');
         $route_match_specification->setValue('one-two-three');
         $decorator = $this->getDecorator($route_match_specification);
-        // IT'S OPPOSITE DAY!!!
-        $this->assertTrue($decorator->isMatchingRoute());
+        $this->assertFalse($decorator->isMatchingRoute());
         //  testing 1-2-1-2 request and route
         $route_match_specification = $this->getRouteMatchSpecification(array('testing' => 'one-two-one-two'));
         $route_match_specification->setParam('testing');
         $route_match_specification->setValue('one-two-one-two');
         $decorator = $this->getDecorator($route_match_specification);
-        // IT'S OPPOSITE DAY!!!
-        $this->assertFalse($decorator->isMatchingRoute());
+        $this->assertTrue($decorator->isMatchingRoute());
     }
 }
