+++ tests/testcases/core/domain/entities/routing/specifications/admin/EspressoEventEditorAddNewTest.php
@@ -2,45 +2,43 @@
 
 namespace EventEspresso\tests\testcases\core\domain\entities\routing\specifications\admin;
 
-use EventEspresso\core\domain\entities\routing\specifications\admin\EspressoEventsListTable;
+use EventEspresso\core\domain\entities\routing\specifications\admin\EspressoEventEditorAddNew;
+use EventEspresso\core\exceptions\InvalidEntityException;
 use EventEspresso\tests\testcases\core\domain\entities\routing\specifications\RouteMatchSpecificationTestBase;
+use PHPUnit\Framework\AssertionFailedError;
 
 /**
- * EspressoEventsListTableTest
+ * EspressoEventEditorAddNewTest
  *
  * @package EventEspresso\core\domain\entities\routing\specifications
  * @author  Brent Christensen
  * @since   4.9.71.p
  */
-class EspressoEventsListTableTest extends RouteMatchSpecificationTestBase
+class EspressoEventEditorAddNewTest extends RouteMatchSpecificationTestBase
 {
 
     /**
      * @param array $request_params
      * @since 4.9.71.p
-     * @return EspressoEventsListTable
-     * @throws \EventEspresso\core\exceptions\InvalidEntityException
+     * @return EspressoEventEditorAddNew
+     * @throws InvalidEntityException
      */
     public function getMultiRouteSpecification(array $request_params)
     {
-        return new EspressoEventsListTable($this->getRequest($request_params));
+        return new EspressoEventEditorAddNew($this->getRequest($request_params));
     }
 
     /**
      * @since 4.9.71.p
-     * @throws \PHPUnit\Framework\AssertionFailedError
-     * @throws \EventEspresso\core\exceptions\InvalidEntityException
+     * @throws AssertionFailedError
+     * @throws InvalidEntityException
      */
     public function testIsMatchingRoute()
     {
         $route_match_specification = $this->getMultiRouteSpecification(
-            array('page' => 'espresso_events')
-        );
-        $this->assertTrue($route_match_specification->isMatchingRoute());
-        $route_match_specification = $this->getMultiRouteSpecification(
             array(
                 'page' => 'espresso_events',
-                'action' => 'default',
+                'action' => 'create_new',
             )
         );
         $this->assertTrue($route_match_specification->isMatchingRoute());
