+++ tests/testcases/core/domain/entities/routing/specifications/admin/EspressoEventEditorEditTest.php
@@ -3,7 +3,9 @@
 namespace EventEspresso\tests\testcases\core\domain\entities\routing\specifications\admin;
 
 use EventEspresso\core\domain\entities\routing\specifications\admin\EspressoEventEditorEdit;
+use EventEspresso\core\exceptions\InvalidEntityException;
 use EventEspresso\tests\testcases\core\domain\entities\routing\specifications\RouteMatchSpecificationTestBase;
+use PHPUnit\Framework\AssertionFailedError;
 
 /**
  * EspressoEventEditorEditTest
@@ -19,7 +21,7 @@ class EspressoEventEditorEditTest extends RouteMatchSpecificationTestBase
      * @param array $request_params
      * @since 4.9.71.p
      * @return EspressoEventEditorEdit
-     * @throws \EventEspresso\core\exceptions\InvalidEntityException
+     * @throws InvalidEntityException
      */
     public function getMultiRouteSpecification(array $request_params)
     {
@@ -28,8 +30,8 @@ class EspressoEventEditorEditTest extends RouteMatchSpecificationTestBase
 
     /**
      * @since 4.9.71.p
-     * @throws \PHPUnit\Framework\AssertionFailedError
-     * @throws \EventEspresso\core\exceptions\InvalidEntityException
+     * @throws AssertionFailedError
+     * @throws InvalidEntityException
      */
     public function testIsMatchingRoute()
     {
