+++ tests/testcases/core/domain/entities/routing/specifications/admin/WordPressPostsEditorEditTest.php
@@ -3,7 +3,9 @@
 namespace EventEspresso\tests\testcases\core\domain\entities\routing\specifications\admin;
 
 use EventEspresso\core\domain\entities\routing\specifications\admin\WordPressPostsEditorEdit;
+use EventEspresso\core\exceptions\InvalidEntityException;
 use EventEspresso\tests\testcases\core\domain\entities\routing\specifications\RouteMatchSpecificationTestBase;
+use PHPUnit\Framework\AssertionFailedError;
 
 /**
  * WordPressPostsEditorEditTest
@@ -17,10 +19,10 @@ class WordPressPostsEditorEditTest extends RouteMatchSpecificationTestBase
 
     /**
      * @param array $request_params
+     * @param array $post_params
      * @param array $server
-     * @since 4.9.71.p
      * @return WordPressPostsEditorEdit
-     * @throws \EventEspresso\core\exceptions\InvalidEntityException
+     * @since 4.9.71.p
      */
     public function getMultiRouteSpecification(array $request_params, array $post_params, array $server)
     {
@@ -31,8 +33,8 @@ class WordPressPostsEditorEditTest extends RouteMatchSpecificationTestBase
 
     /**
      * @since 4.9.71.p
-     * @throws \PHPUnit\Framework\AssertionFailedError
-     * @throws \EventEspresso\core\exceptions\InvalidEntityException
+     * @throws AssertionFailedError
+     * @throws InvalidEntityException
      */
     public function testIsMatchingRoute()
     {
