+++ tests/testcases/core/domain/entities/routing/specifications/admin/WordPressPostsListTableTest.php
@@ -3,7 +3,9 @@
 namespace EventEspresso\tests\testcases\core\domain\entities\routing\specifications\admin;
 
 use EventEspresso\core\domain\entities\routing\specifications\admin\WordPressPostsListTable;
+use EventEspresso\core\exceptions\InvalidEntityException;
 use EventEspresso\tests\testcases\core\domain\entities\routing\specifications\RouteMatchSpecificationTestBase;
+use PHPUnit\Framework\AssertionFailedError;
 
 /**
  * WordPressPostsListTableTest
@@ -30,8 +32,8 @@ class WordPressPostsListTableTest extends RouteMatchSpecificationTestBase
 
     /**
      * @since 4.9.71.p
-     * @throws \PHPUnit\Framework\AssertionFailedError
-     * @throws \EventEspresso\core\exceptions\InvalidEntityException
+     * @throws AssertionFailedError
+     * @throws InvalidEntityException
      */
     public function testIsMatchingRoute()
     {
